requirements.txt
pandas
numpy
pillow
jupyterlab
